oled: WIP:
This design doesn't work so nicely because the data left in the FIFO
depends on whether our boy failed or not.  We should probably just take
the address in the FIFO too and then we know that always gets consumed,
and then whether we push any more in can depend on whether we got
ACK/NACk.