rom: try SequenceBreak idea
It doesn't work because OLED doesn't actually know what state the
device's stream parser is in; it doesn't know whether an F0 (or any
byte) is at a control byte position (and therefore a seq break) or part
of data.  *Any* byte can be part of data, so none is available for use
this way.

We could add some knowledge of control bytes and continuation to OLED..
or we could not. :)