oled: clean up bus usage, extract chpr
Now that we have a dedicated character printer circuit that handles
locating and wrapping, it can be used by more commands than just PRINT.
(In this case, ID.)

I want to make this a submodule, but I don't really know how to deal
with simultaneous/arbitrating access to ROMWriter/Locator/etc. -- I had
a few false starts before just making the "chpr" circuit simultaneous
within OLED.

Also played around with different things to see what was more cell
efficient.  e.g.

    with m.If(x != 0):
        m.d.sync += [
            self.x.eq(x),
            self.some_submod.x.eq(x),
        ]
    with m.Else():
        m.d.sync += self.some_submod.x.eq(0)

I got smart and thought I'd change it to this:

    m.d.sync += [
        self.x.eq(Mux(x != 0, x, self.x)),
        self.some_submod.x.eq(x),
    ]

That increased cells.  So I thought maybe the "needless reassignment" in
the self.x case was adding to it, and tried this:

    with m.If(x != 0):
        m.d.sync += self.x.eq(x)
    m.d.sync += self.some_submod.x.eq(x)

To my surprise, this was still worse than the original.  I'm realising
it's likely because this unconditionally requires passing x through to
self.some_submod.x, whereas when it's nested in the If, that only
happens when x != 0; otherwise we pass through 0 and don't actually need
to pull that line in again.  Or something?

I think I also had a moment of realising that flow type-style folding
isn't a thing:  "with m.If(x == 0): m.d.sync += self.x.eq(x)" is more
expensive than "with m.If(x == 0): m.d.sync += self.x.eq(0)", because
the latter is still making extra wires to x even though we know it'll
assign 0.

Anyway, all the specific examples above might be completely wrong since
it's been an hour or so, but point is, gotta check these things.